Fix issue #41 – Project Fully Refactored, Files Renamed, CONTRIBUTING.md added
