+++ CONTRIBUTING.md
@@ -1,0 +1,25 @@
+# **How to Contribute**
+
+Look through Collection_of_Algorithms folder. If you know of a certain C++ algorithm or a problem, however you notice that it is missing from our collection or if it exists but can be improved:
+* proceed to file an issue, stating what you want to work on
+* create a file (following our naming conventions) to work on new code or edit existing code in existing files
+* create a pull request to merge your work
+* look out for comments on your pull request that request changes or additions to your code
+
+## **File Naming Convention**
+
+### All new files should follow the following naming format:
+
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; **MergeSort.cpp**
+
+* Capitalize the first letters of every word in the title
+* No spaces or special characters between words
+* The rest of the letters are lowercase 
+
+### Do not name files in the following format:
+
+* mergeSort.cpp
+* merge_sort.cpp
+* merge sort.cpp
+* Merge-Sort.cpp
+* ...
