+++ Collection_of_Algorithms/Backtracking/NQueens.cpp
