+++ Collection_of_Algorithms/Mathematics/BinaryShuffle.cpp
