+++ Collection_of_Algorithms/Mathematics/BoringFactorials.cpp
