+++ Collection_of_Algorithms/Mathematics/CubeFree.cpp
