+++ Collection_of_Algorithms/Mathematics/DivisibleBy41.cpp
