+++ Collection_of_Algorithms/Mathematics/EulerTotientFunction.cpp
