+++ Collection_of_Algorithms/Mathematics/FibonacciSum.cpp
