+++ Collection_of_Algorithms/Mathematics/FindTheGoodSets.cpp
