+++ Collection_of_Algorithms/Mathematics/ModularExp.cpp
