+++ Collection_of_Algorithms/Mathematics/SEGsieve.cpp
