+++ Collection_of_Algorithms/Mathematics/SubsetSumProblem.c
