+++ Collection_of_Algorithms/Other/WordCount.cpp
