+++ Collection_of_Algorithms/Recursion/ChessRecursion.cpp
