+++ Collection_of_Algorithms/Trees/StaticTree.cpp
