--- algos/SudokuSolve.cpp
@@ -1,123 +1,0 @@
-// A Backtracking program  in C++ to solve Sudoku problem
-#include <iostream>
-#define UNASSIGNED 0
-#define N 4
-#define SQN 2
-using namespace std;
-// This function finds an entry in grid that is still unassigned
-bool Find0Loc(int grid[N][N], int &row, int &col);
-
-// Checks whether it will be legal to assign num to the given row,col
-bool isSafe(int grid[N][N], int row, int col, int num);
-
-/* Takes a partially filled-in grid and attempts to assign values to
-  all unassigned locations in such a way to meet the requirements
-  for Sudoku solution (non-duplication across rows, columns, and boxes) */
-bool SolveSudoku(int grid[N][N])
-{
-    int row, col;
-    if (!Find0Loc(grid, row, col))
-       return true; // success!
-
-    // consider digits 1 to 9
-    for (int num = 1; num <= N; num++)
-    {
-        // if looks promising
-        if (isSafe(grid, row, col, num))
-        {
-            // make tentative assignment
-            grid[row][col] = num;
-            if (SolveSudoku(grid))
-                return true; //Done!
-            // failure, unmake & try again
-            grid[row][col] = UNASSIGNED;
-        }
-    }
-    return false; // this triggers backtracking
-}
-
-/* Searches the grid to find an entry that is still unassigned. If
-   found, the reference parameters row, col will be set the location
-   that is unassigned, and true is returned. If no unassigned entries
-   remain, false is returned. */
-bool Find0Loc(int grid[N][N], int &row, int &col)
-{
-    for (row = 0; row < N; row++)
-        for (col = 0; col < N; col++)
-            if (grid[row][col] == UNASSIGNED)
-                return true;
-    return false;
-}
-
-/* Returns a boolean which indicates whether any assigned entry
-   in the specified row matches the given number. */
-bool UsedInRow(int grid[N][N], int row, int num)
-{
-    for (int col = 0; col < N; col++)
-        if (grid[row][col] == num)
-            return true;
-    return false;
-}
-
-/* Returns a boolean which indicates whether any assigned entry
-   in the specified column matches the given number. */
-bool UsedInCol(int grid[N][N], int col, int num)
-{
-    for (int row = 0; row < N; row++)
-        if (grid[row][col] == num)
-            return true;
-    return false;
-}
-
-/* Returns a boolean which indicates whether any assigned entry
-   within the specified 3x3 box matches the given number. */
-bool UsedInBox(int grid[N][N], int boxStartRow, int boxStartCol, int num)
-{
-    for (int row = 0; row < SQN; row++)
-        for (int col = 0; col < SQN; col++)
-            if (grid[row+boxStartRow][col+boxStartCol] == num)
-                return true;
-    return false;
-}
-
-/* Returns a boolean which indicates whether it will be legal to assign
-   num to the given row,col location. */
-bool isSafe(int grid[N][N], int row, int col, int num)
-{
-    /* Check if 'num' is not already placed in current row,
-       current column and current 3x3 box */
-    return !UsedInRow(grid, row, num) &&
-           !UsedInCol(grid, col, num) &&
-           !UsedInBox(grid, row - row%SQN , col - col%SQN, num);
-}
-
-/* A utility function to print grid  */
-void printGrid(int grid[N][N])
-{
-    for (int row = 0; row < N; row++)
-    {
-       for (int col = 0; col < N; col++)
-             printf("%2d", grid[row][col]);
-        printf("\n");
-    }
-}
-
-/* Driver Program to test above functions */
-int main()
-{
-    // 0 means unassigned cells
-    int grid[N][N];
-    for(int i =0; i<N; i++){
-        for(int j=0; j<N; j++){
-            cout<<i<<j;
-            cin>>grid[i][j];
-        }
-    }
-
-    if (SolveSudoku(grid) == true)
-          printGrid(grid);
-    else
-         printf("No solution exists");
-
-    return 0;
-}
